Moves Boards module to Tictactoe::Boards
