+++ spec/tictactoe/state_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 require 'tictactoe/state'
-require 'boards/three_by_three_board'
+require 'tictactoe/boards/three_by_three_board'
 
 RSpec.describe Tictactoe::State do
   def look_at(state, location)
@@ -28,7 +28,7 @@ RSpec.describe Tictactoe::State do
 
   describe "given a 3x3 board" do
     before(:each) do
-      @board = Boards::ThreeByThreeBoard.new
+      @board = Tictactoe::Boards::ThreeByThreeBoard.new
       @state = described_class.new(@board)
     end
 
@@ -77,7 +77,7 @@ RSpec.describe Tictactoe::State do
       end
     end
 
-    Boards::ThreeByThreeBoard.new.lines.each do |line|
+    Tictactoe::Boards::ThreeByThreeBoard.new.lines.each do |line|
       describe "with a line for player a" do
         before(:each) do
           line.each do |l|
